move cache auto continue and retry to all api backend
